// frontend/src/openapi/models/message.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ReadersLounge API
 * ReadersLounge API
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 * @interface Message
 */
export interface Message {
  /**
   *
   * @type {number}
   * @memberof Message
   */
  message_id: number;
  /**
   *
   * @type {number}
   * @memberof Message
   */
  user_id: number;
  /**
   *
   * @type {string}
   * @memberof Message
   */
  content: string;
  /**
   *
   * @type {string}
   * @memberof Message
   */
  sent_at: string;
}