// frontend/src/openapi/models/send-message-req-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ReadersLounge API
 * ReadersLounge API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SendMessageReqBody
 */
export interface SendMessageReqBody {
    /**
     * 
     * @type {string}
     * @memberof SendMessageReqBody
     */
    'content': string;
    /**
     * 
     * @type {number}
     * @memberof SendMessageReqBody
     */
    'room_id': number;
}